NEW DBField validation
