+++ src/Forms/EmailField.php
@@ -2,6 +2,9 @@
 
 namespace SilverStripe\Forms;
 
+use SilverStripe\Core\Validation\ConstraintValidator;
+use Symfony\Component\Validator\Constraints;
+
 /**
  * Text input field with validation for correct email format according to RFC 2822.
  */
@@ -18,32 +21,27 @@ class EmailField extends TextField
     }
 
     /**
-     * Validates for RFC 2822 compliant email addresses.
-     *
-     * @see http://www.regular-expressions.info/email.html
-     * @see http://www.ietf.org/rfc/rfc2822.txt
-     *
      * @param Validator $validator
      *
      * @return string
      */
     public function validate($validator)
     {
-        $result = true;
         $this->value = trim($this->value ?? '');
 
-        $pattern = '^[a-z0-9!#$%&\'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&\'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$';
+        $result = true;
+        // $message = _t('SilverStripe\\Forms\\EmailField.VALIDATION', 'Please enter an email address');
+        // $validationResult = ConstraintValidator::validate(
+        //     $this->value,
+        //     new Constraints\Email(message: $message)
+        // );
+        $validationResult = $this->getModelField()->validate();
 
-        // Escape delimiter characters.
-        $safePattern = str_replace('/', '\\/', $pattern ?? '');
-
-        if ($this->value && !preg_match('/' . $safePattern . '/i', $this->value ?? '')) {
+        if (!$validationResult->isValid()) {
             $validator->validationError(
                 $this->name,
-                _t('SilverStripe\\Forms\\EmailField.VALIDATION', 'Please enter an email address'),
-                'validation'
+                $validationResult->getMessages()[0]['message'],
             );
-
             $result = false;
         }
