+++ src/Forms/FormRequestHandler.php
@@ -216,10 +216,10 @@ class FormRequestHandler extends RequestHandler
         // Action handlers may throw ValidationExceptions.
         try {
             // Or we can use the Validator attached to the form
-            $result = $this->form->validationResult();
-            if (!$result->isValid()) {
-                return $this->getValidationErrorResponse($result);
-            }
+            // $result = $this->form->validationResult();
+            // if (!$result->isValid()) {
+            //     return $this->getValidationErrorResponse($result);
+            // }
 
             // First, try a handler method on the controller (has been checked for allowed_actions above already)
             $controller = $this->form->getController();
