+++ src/ORM/UnsavedRelationList.php
@@ -6,6 +6,7 @@ use InvalidArgumentException;
 use ArrayIterator;
 use SilverStripe\Model\List\ArrayList;
 use SilverStripe\ORM\FieldType\DBField;
+use SilverStripe\ORM\FieldType\DBFieldHelper;
 use Traversable;
 
 /**
